bot/requirements.txt
misskey.py
requests
